Updated is_valid() and created get_coord()
I have updated is_valid() so that it returns a list of invalid squares
instead of a bool. To achieve this I made a function that gets the square information ('A1') and transforms it to the coordinate position on the 9x9 grid (0,0). I have tested this and it works, will now add this functionality to the user interface.